ga_opt/agent/requirements.txt
openai 
python-dotenv
openai-agents